card_file_box: Back-End: Simplify MongoDB insert and delete operations.
The `insert` method in the `MongoDB` class has been simplified to use the `insertMany` function of the MongoDB driver. This change improves the efficiency of inserting multiple documents into a collection.

The `delete` method in the `MongoDB` class has also been simplified to use the `deleteOne` function of the MongoDB driver. This change improves the efficiency of deleting a single document from a collection.

These changes were made to optimize the MongoDB operations and improve the overall performance of the back-end application.