✨ feat(db): Add support for exporting MongoDB databases in different formats
The `dump` method in the `MongoDB` class now supports exporting MongoDB databases in different formats like JSON and BSON. The `exportType` parameter determines the format of the export.

WHY: This change was made to provide flexibility and enable users to export MongoDB databases in formats that suit their needs, such as JSON and BSON.

🌐 Updated `README.md`

✨ feat(ui): Display loading spinner while data is being loaded

Added a loading spinner and a text message to indicate that the requested database is being loaded. This provides better visual feedback to the user and improves the overall user experience.

WHY: It is important to provide visual feedback to the user during data loading to indicate that the system is working and to prevent users from assuming that the application is unresponsive.

🔧 chore(ui): Minor CSS modifications

Removed the transition effect on the background color of the `#home` element in the CSS file, as it is no longer needed.

🔧 chore(ui): Disabled data inclusion toggle for non-SQL servers

The `includeData` toggle is now disabled for non-SQL servers. This prevents users from including data in the export process for non-SQL servers, as they do not support including data.

WHY: The `includeData` toggle should only be enabled for SQL servers, as they are the only ones that support including data in the export process.

✨ feat(ui): Dynamic file types for export format selection

The export format selection dropdown now displays options based on the supported file types of the selected server. The file types are fetched from the `driver` object of the server and displayed in the dropdown.

WHY: This change was made to dynamically display the supported file types for the export format selection, providing a better user experience and preventing users from selecting unsupported file types.

🐛 fix(ui): Enable data inclusion toggle for SQL servers