✨🐛 **Fix: Update docker command in .gitlab-ci.yml**
- Change the command to include the flag "-e" instead of "--exit-code" when running the "docker scout cves" command.
- Also specify the severity level as "critical" using the flag "--only-severity critical".

These changes were made in order to ensure that the Docker image is scanned for Critical Common Vulnerabilities and Exposures (CVEs) during the CI/CD pipeline. The previous command was using an incorrect flag, resulting in the scan not being performed properly. By using the correct flag and specifying the severity level, we can ensure that critical vulnerabilities are identified and addressed.