rocket: chore(ci): Update .gitlab-ci.yml
The `.gitlab-ci.yml` file was updated to fix a bug in the CI pipeline. The line `docker scout cves $IMAGE_PATH -e --only-severity critical` was modified to place double quotes around the `$IMAGE_PATH` variable to ensure proper parsing. This change allows the "docker scout cves" command to run with the correct image path argument.

This commit improves the reliability of the CI pipeline by fixing a bug in the `docker scout cves` command argument.