construction_worker: Chore: Update .gitlab-ci.yml and README.md
- Updated the `.gitlab-ci.yml` file.
   - Commented out the command to install curl and scout-cli.
   - Removed the command to check for critical CVEs using scout-cli.
- Updated the `README.md` file.
   - Added the command to check for CVEs using scout-cli.

The changes were made to temporarily disable the installation of curl and scout-cli and the checking of critical CVEs in the pipeline due to some issues. The README.md was updated to include the command for checking CVEs using scout-cli. These changes are necessary for the smooth functioning of the pipeline and to ensure the security of the codebase.