fire: Remove unused code
The `container.component.html` file had an unused `<a>` element related to a disclaimer. This element was removed from the file.

:recycle: Refactor button text
In the `servers.component.html` file, the text on a button was changed from "Guess" to "Guess Credentials" to clarify its purpose.

:art: Update styling in servers.component.scss
The `servers.component.scss` file was updated to include a new styling rule. The height of the container was adjusted to be equal to the viewport height minus 122 pixels, and an overflow rule was added to enable scrolling when necessary.

:bug: Fix loading server subscription
In the `servers.component.ts` file, the subscription to the `loadingServer` observable was updated to correctly increment the `loading` value when a loading event occurs.

:hammer_and_wrench: Improve table duplication feature
The `advanced.component.html` file was modified to update the tooltip text on a button. The button is used to duplicate a table, and the tooltip was changed to inform users that the duplication process will not duplicate references or indexes on the new table.