✨ front: Update container component HTML and SCSS files
- Removed commented out code for loading spinner and error message in the
  container.component.html file.
- Replaced the text "Loading Servers ..." with "Scanning Connections ..."
  in the servers.component.html file.

The changes were made to improve the clarity and readability of the code. The
commented out code in the container component HTML file was removed as it was
no longer needed. Additionally, the text in the servers component HTML file
was updated to provide a more accurate description of the loading process.