memo: **docs**: Remove unnecessary lines from README.md
The changes were made to remove unnecessary lines from the README.md file. This was done to improve the readability and organization of the document.

:recycle: **refactor**: Change routes in core-routing.module.ts

The routes in the `core-routing.module.ts` file were modified. A new route with the path `query/:code` was added, which maps to the `QueryComponent`. This allows for dynamic routing based on a code parameter.

:sparkles: **feat**: Add code to QueryComponent

Code was added to the `QueryComponent` to handle the `code` parameter in the route. If the `code` parameter is present, it sets the `replayQuery` property of the component to the value of the `code` parameter.

:recycle: **refactor**: Update button click event in history.component.html

The click event of a button in the `history.component.html` file was updated to call the `goToQuery` method in the component with the `his.query` parameter. This change was necessary to properly handle the button click event.

:recycle: **refactor**: Update goToQuery method in history.component.ts

The `goToQuery` method in the `history.component.ts` file was updated to use the `Router` service to navigate to a new route based on the selected server, database, and table. The `toggle` method of the `DrawerService` is also called to close the drawer after navigation.

:recycle: **refactor**: Move code component HTML content to separate file

The HTML content of the `code.component.html` file was moved to a separate file to improve the organization and readability of the code.

:recycle: **refactor**: Update code component styles

The styles of the `code.component.scss` file were updated to improve the layout and appearance of the component.

:recycle: **refactor**: Update styles and remove unnecessary lines in styles.scss

The styles in the `styles.scss` file were updated to improve the overall appearance and organization of the styles. Unnecessary lines were also removed to clean up the file.