art: Chore: Update README.md and query.component.scss
In this commit, the README.md file was updated to include new tasks and remove unnecessary ones. The update includes adding a new type to autocomplete and right-column, modifying default to function suggestions, and discussing the mandatory size in postgre and export in mongo.

The query.component.scss file was also modified, where the width of ngx-monaco-editor was set to 50% and the first child was resized and had horizontal overflow.

WHY: The changes were made to improve the functionality and appearance of the application.