sparkles: Add create issue button to app.component.ts
Added a new button to the app.component's template that links to the GitLab project's issue creation page. This allows users to easily create new issues for bug reports or feature requests. The button is positioned in the bottom right corner of the screen.

:art: Remove unused import in app.module.ts

Removed the unused import of `ConnectionInfoDialog` from the list of imports in app.module.ts.

:fire: Remove connection-dialog.html

Deleted the connection-dialog.html file as it is no longer needed. It was previously used as a template for the connection info dialog, but the dialog has been removed from the app.

:recycle: Refactor container.component.html

Refactored the container.component.html file to remove unnecessary code related to the connection info dialog. It also removed some unused mat-menu items.

:recycle: Refactor container.component.ts

Refactored the container.component.ts file to remove the unused `showConnection()` method and the `ConnectionInfoDialog` component.

:sparkles: Add advanced settings editor to advanced.component.html

Added a new ngx-monaco-editor component to the advanced.component.html file. This editor will be used to display advanced settings in JSON format. It is styled with a height of 30vh.

:sparkles: Add JSON editor options to advanced.component.ts

Added options for the JSON editor in the advanced.component.ts file. The editor will display JSON content and is set to read-only mode.

These changes were made to improve the user experience and remove unused code from the application. The create issue button allows users to easily report bugs or request new features. The removal of unused code improves maintainability and reduces complexity. The addition of the JSON editor in advanced.component.html provides a better way to display advanced settings.