✨ :lipstick: Updated Angular Material imports in app.module.ts
The commit updates the imports in the `app.module.ts` file to include the `MatTabsModule` from Angular Material. These changes were made in order to use Angular Material tabs in the application.

🛠️ :hammer: Refactored container.component.html

The commit refactors the `container.component.html` file. The changes include:
- Separate the server host and port from the server/database name
- Add a line break between the server host/port and the server/database name

The changes were made to improve the readability and layout of the server information displayed in the application.

🔧 :wrench: Updated container.component.scss

The commit updates the `container.component.scss` file. The changes include:
- Adjusting the line height, font weight, and font size of the server information in the drawer component
- Removing the border at the bottom of the panel
- Adjusting the height of the tab links

These changes were made to improve the visual appearance and styling of the container component in the application.

🔄 :arrows_counterclockwise: Refactored container.component.ts

The commit refactors the `container.component.ts` file. The changes include:
- Removing the unused import `MAT_DIALOG_DATA`
- Adding the `environment` variable for better code organization

These changes were made to clean up the code and improve the organization of dependencies in the container component.

🎨 :art: Updated styles.scss

The commit updates the `styles.scss` file. The changes include:
- Adding a backdrop filter blur to the material tooltips

These changes were made to improve the visual appearance of material tooltips in the application.