🎨 Add styles to container component
Added styles to the container component to adjust the position of the ink bar in the mat-drawer.

💡 Improve logic in advanced.component.ts

Improved the logic in the advanced component by changing the index of accessing the Tabs array.

🐛 Fix dialog data in explore.component.ts

Modified the dialog data in the explore component to pass an object with the row and updateInPlace properties.

✨ Add close button to iframe-dialog.html

Added a close button to the iframe dialog for closing the dialog.

🎨 Update styles in insert.component.scss

Updated the styles in the insert component to adjust the margin of the result class.

🐛 Fix paginator initialization in insert.component.ts

Fixed the paginator initialization in the insert component by moving it to the ngOnInit method.

💄 Update toolbar buttons in insert.component.html

Updated the toolbar buttons in the insert component to use mat-flat-button and mat-stroked-button.

🐛 Fix row generation in insert.component.ts

Fixed the row generation in the insert component by only generating columns that are not the action column.

🐛 Fix matRowDef and matHeaderRowDef in insert.component.html

Fixed the matRowDef and matHeaderRowDef directives in the insert component by using the displayedColumns property without a selector.

🚚 Remove unused import in tables.component.ts

Removed the unused import in the tables component.

♻️ Refactor Tabs constant in tables.component.ts

Refactored the Tabs constant in the tables component to be an array of strings instead of objects.

🔥 Remove unused column.ts file

Removed the unused column.ts file.

✨ Add updateInPlace property to UpdateDataDialogComponent

Added the updateInPlace property to the UpdateDataDialogComponent to determine if the update should be done in place or not.

🐛 Fix update request in UpdateDataDialogComponent

Fixed the update request in the UpdateDataDialogComponent to only make the request if the updateInPlace property is true.

🎨 Update styles in styles.scss

Updated the styles in styles.scss to set the border-spacing of tables to 0px.