lipstick: Update UI in diagram.component.html and update-data-dialog.component.html
The UI in the diagram.component.html and update-data-dialog.component.html files was updated. The "Hide columns without tags" tooltip in the diagram component was modified to remove the mention of an emoji. In the update-data-dialog component, the label for each column suggestion was simplified to just display the column key.