🚀 [front]feat: Add connection dialog styling changes
- Modified the style of the `ngx-monaco-editor` in the `add-connection-dialog.html` file to have a width of 100%.

This commit improves the styling of the connection dialog to make it more visually appealing and ensure that the editor takes up the entire width of the dialog.

🚀 [front]chore: Remove unnecessary style in connection dialog

- Removed the `style` attribute in the `connection-dialog.html` file that set the height and width of the `ngx-monaco-editor`.

This commit removes unnecessary inline styles in the connection dialog to allow the editor to take up its default height and width.

🚀 [front]fix: Redirect to query route properly

- Updated the `redirectTo` path in the `core-routing.module.ts` file to include a trailing slash.

This commit fixes a bug where the redirect to the query route was not working properly. The addition of the trailing slash ensures that the redirect path is correct.

🚀 [front]style: Update explore component UI

- Made changes to the UI of the explore component in the `explore.component.html` file.
- Added a button with a menu for performing actions on selected rows.
- Adjusted the layout and styling of the table.

This commit updates the UI of the explore component to improve the user experience and enhance the visual appeal of the table.

🚀 [front]chore: Remove unnecessary style in export query dialog

- Removed the `style` attribute in the `export-query-dialog.html` file that set the width and height of the `ngx-monaco-editor`.

This commit removes unnecessary inline styles in the export query dialog to allow the editor to take up its default width and height.

🚀 [front]chore: Remove unnecessary style in export result dialog

- Removed the `style` attribute in the `export-result-dialog.html` file that set the width and height of the `ngx-monaco-editor`.

This commit removes unnecessary inline styles in the export result dialog to allow the editor to take up its default width and height.

🚀 [front]style: Update query component UI

- Modified the style of the `ngx-monaco-editor` in the `query.component.html` file to adjust the height.

This commit updates the UI of the query component to improve the layout and ensure that the editor has a suitable height.

🚀 [front]style: Update query component styles

- Made changes to the styles of the query component in the `query.component.scss` file.
- Adjusted the width and height of the `ngx-monaco-editor` to create a better layout.

This commit improves the visual appearance of the query component by adjusting the width and height of the editor and applying appropriate styles.

🚀 [front]fix: Check for empty code parameter in query component

- Updated the check for the 'code' parameter in the query component to also verify that the value is not empty after trimming.

This commit ensures that the 'code' parameter in the query component is checked for both existence and non-empty value to prevent errors.

🚀 [front]style: Update load component styles

- Made changes to the styles of the load component in the `load.component.html` file.
- Adjusted the width and height of the `ngx-monaco-editor` to create a better layout.

This commit updates the visual styling of the load component, improving the layout and ensuring that the editor has appropriate dimensions.

🚀 [front]chore: Remove unnecessary style in update data dialog

- Removed the `style` attribute in the `update-data-dialog.component.html` file that set the height and width of the `ngx-monaco-editor`.

This commit removes unnecessary inline styles in the update data dialog to allow the editor to take up its default height and width.

🚀 [front]style: Add global styles

- Added global styles to the `styles.scss` file to customize the appearance of the `ngx-monaco-editor`.

This commit adds global styles to the project's main stylesheet to customize the appearance of the `ngx-monaco-editor` component.

These changes were made to improve the visual appearance, usability, and overall user experience of the application by refining the styling and layout of various components.